list table css problems and PR fixes
