+++ src/assets/styles/main.scss
@@ -194,6 +194,10 @@ html {
   min-height: 30px;
 }
 
+.min-height-40 {
+  min-height: 40px;
+}
+
 .height-50px {
   height: 50px;
   max-height: 50px;
@@ -455,4 +459,15 @@ $color-light: hsl(0, 0%, 96%);
 
 .white-space-pre {
   white-space: pre;
+}
+
+/* editor dropdown width */
+.docs-selection {
+  max-width: 450px;
+  min-width: 300px;
+}
+
+/* editor dropdown width */
+.name-table {
+  max-width: 300px;
 }
